requirements.txt
opencv-contrib-python==4.7.0.72
numpy==1.24.3
networkx==3.1
tqdm==4.66.1
torch>=2.2.0
torchvision>=0.17.0
pillow>=10.0.0
